Use config flag in options
