+++ cmd/ghost/cmd.go
@@ -19,15 +19,16 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/chronicleprotocol/oracle-suite/cmd"
+	"github.com/chronicleprotocol/oracle-suite/pkg/config"
 	"github.com/chronicleprotocol/oracle-suite/pkg/config/ghost"
 	"github.com/chronicleprotocol/oracle-suite/pkg/log/logrus/flag"
 )
 
 type options struct {
 	flag.LoggerFlag
-	ConfigFilePath []string
-	Config         ghost.Config
-	GoferNoRPC     bool
+	config.ConfigFiles
+	Config     ghost.Config
+	GoferNoRPC bool
 }
 
 func NewRootCommand(opts *options) *cobra.Command {
@@ -41,12 +42,7 @@ func NewRootCommand(opts *options) *cobra.Command {
 	}
 
 	rootCmd.PersistentFlags().AddFlagSet(flag.NewLoggerFlagSet(&opts.LoggerFlag))
-	rootCmd.PersistentFlags().StringSliceVarP(
-		&opts.ConfigFilePath,
-		"config", "c",
-		[]string{"./config.hcl"},
-		"ghost config file",
-	)
+	rootCmd.PersistentFlags().AddFlagSet(config.NewConfigFilesFlagSet(&opts.ConfigFiles))
 	rootCmd.PersistentFlags().BoolVar(
 		&opts.GoferNoRPC,
 		"gofer.norpc",
