+++ cmd/spectre/cmd_run.go
@@ -21,8 +21,6 @@ import (
 	"os/signal"
 
 	"github.com/spf13/cobra"
-
-	"github.com/chronicleprotocol/oracle-suite/pkg/config"
 )
 
 func NewRunCmd(opts *options) *cobra.Command {
@@ -31,7 +29,7 @@ func NewRunCmd(opts *options) *cobra.Command {
 		Args:    cobra.ExactArgs(0),
 		Aliases: []string{"agent"},
 		RunE: func(_ *cobra.Command, _ []string) error {
-			if err := config.LoadFiles(&opts.Config, opts.ConfigFilePath); err != nil {
+			if err := opts.LoadConfigFiles(&opts.Config); err != nil {
 				return err
 			}
 			ctx, _ := signal.NotifyContext(context.Background(), os.Interrupt)
