+++ cmd/spire/cmd_agent.go
@@ -1,4 +1,4 @@
-//  Copyright (C) 2020 Maker Ecosystem Growth Holdings, INC.
+//  Copyright (C) 2021-2023 Chronicle Labs, Inc.
 //
 //  This program is free software: you can redistribute it and/or modify
 //  it under the terms of the GNU Affero General Public License as
@@ -21,8 +21,6 @@ import (
 	"os/signal"
 
 	"github.com/spf13/cobra"
-
-	"github.com/chronicleprotocol/oracle-suite/pkg/config"
 )
 
 func NewAgentCmd(opts *options) *cobra.Command {
@@ -32,7 +30,7 @@ func NewAgentCmd(opts *options) *cobra.Command {
 		Args:    cobra.ExactArgs(0),
 		Short:   "Starts the Spire agent",
 		RunE: func(_ *cobra.Command, _ []string) error {
-			if err := config.LoadFiles(&opts.Config, opts.ConfigFilePath); err != nil {
+			if err := opts.LoadConfigFiles(&opts.Config); err != nil {
 				return err
 			}
 			ctx, _ := signal.NotifyContext(context.Background(), os.Interrupt)
