+++ cmd/spire/cmd_pull.go
@@ -49,7 +49,7 @@ func NewPullPriceCmd(opts *options) *cobra.Command {
 		Args:  cobra.ExactArgs(2),
 		Short: "Pulls latest price for a given pair and feed",
 		RunE: func(_ *cobra.Command, args []string) error {
-			if err := config.LoadFiles(&opts.Config, opts.ConfigFilePath); err != nil {
+			if err := opts.LoadConfigFiles(&opts.Config); err != nil {
 				return err
 			}
 			ctx, ctxCancel := signal.NotifyContext(context.Background(), os.Interrupt)
